+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format
 is roughly based on [Keep a Changelog], and this project tries to adheres to
 [Semantic Versioning].
 
+## [0.0.31] - TBD
+
+### Added
+
+- Added a by-operation iterator for alignment states (`CigletOpIter`)
+- Added `max_inc` field to `CigletSpecs` (behind `fuzzing` feature gate)
+
 ## [0.0.30] - 2026-06-24
 
 ### Added
@@ -11,7 +18,6 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `coordinate_sort` for performing coordinate sort on `SamData` records 
 - Added conversion from `ByteSpecs` into `AminoAcidSpecs` and `NucleotideSpecs` (behind `fuzzing` feature gate)
 - Added `display_stack` for displaying an error and its stack (similar to `OrFail` without aborting)
-- Added `max_inc` field to `CigletSpecs` (behind `fuzzing` feature gate)
 
 ### Changed
 
@@ -556,6 +562,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - **Added**: Initial internal release. Provides various readers and types for bioinformatics data manipulation.
 
 <!-- Versions -->
+[0.0.31]: https://github.com/CDCgov/zoe/compare/v0.0.30...v0.0.31
 [0.0.30]: https://github.com/CDCgov/zoe/compare/v0.0.29...v0.0.30
 [0.0.29]: https://github.com/CDCgov/zoe/compare/v0.0.28...v0.0.29
 [0.0.28]: https://github.com/CDCgov/zoe/compare/v0.0.27...v0.0.28